# docs/api/openapi.yaml
openapi: '3.1.1'
info:
  title: A365 Gateway API
  description: 'Connect independently hosted agents to Microsoft Entra, Agent 365, Purview and Prompt Shields. Administration uses an Entra access token; agent ingress uses a gateway-issued secret. Both use Authorization: Bearer, but are not interchangeable. Evaluate → call your model only on allow → submit the interaction. HTTP 202 means queued, never confirmed delivery. All timestamps are UTC. Interactive requests act on the live deployment.'
  version: v1
servers:
  - url: http://127.0.0.1:5080/
paths:
...
  /api/v1/agents:
    post:
      tags:
        - Agents
      summary: Register an external agent
      description: "Creates a registration and queues provisioning. Save the returned one-time credential securely. A 202 response is not Active. If the response is lost, reconcile using the exact externalAgentId instead of blindly creating another registration.\n\nAuthorization policy: AdministratorOnly."
      operationId: Agents.RegisterAgent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
        required: true
      responses:
        '202':
          description: Accepted
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RegisterAgentResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAgentResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RegisterAgentResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Unprocessable Entity
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '502':
          description: Bad Gateway
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
    get:
      tags:
        - Agents
      summary: Search and paginate agents
      description: "Follow nextCursor until absent. Search and pagination are server-side; do not interpret one page as the complete inventory.\n\nAuthorization policy: AllControlPlane."
      operationId: Agents.ListAgents
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: environment
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            pattern: '^-?(?:0|[1-9]\d*)$'
            type:
              - integer
              - string
            format: int32
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
  '/api/v1/agents/{agentId}':
    get:
      tags:
        - Agents
      summary: Read an agent
      description: "Returns registration state, protection configuration, provisioning references and rowVersion. Active registration does not itself establish protection readiness.\n\nAuthorization policy: AllControlPlane."
      operationId: Agents.GetAgent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AgentDetailDto'
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetailDto'
            text/json:
              schema:
                $ref: '#/components/schemas/AgentDetailDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
    delete:
      tags:
        - Agents
      summary: Delete a gateway registration
      description: "Deletes the gateway registration according to lifecycle rules. Linked Microsoft identities and Registry resources are not automatically deleted.\n\nAuthorization policy: AdministratorOnly."
      operationId: Agents.DeleteAgent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Accepted
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/DeleteAgentResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAgentResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/DeleteAgentResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
  '/api/v1/agents/{agentId}/credentials':
    get:
      tags:
        - Agents
      summary: List agent credential metadata
      description: "Returns metadata only. Existing secret values cannot be retrieved.\n\nAuthorization policy: AdministratorOnly."
      operationId: Agents.ListAgentIngressCredentials
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AgentIngressCredentialListResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AgentIngressCredentialListResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AgentIngressCredentialListResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
    post:
      tags:
        - Agents
      summary: Issue an agent credential
      description: "Returns a new secret once with no-store caching. Save it securely and do not automatically retry an uncertain issuance.\n\nAuthorization policy: AdministratorOnly."
      operationId: Agents.IssueAgentIngressCredential
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/IssueAgentIngressCredentialResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAgentIngressCredentialResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/IssueAgentIngressCredentialResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
  '/api/v1/agents/{agentId}/credentials/{credentialId}':
    delete:
      tags:
        - Agents
      summary: Revoke an agent credential
      description: "Revokes only the selected credential. The server enforces the last-usable-credential guard.\n\nAuthorization policy: AdministratorOnly."
      operationId: Agents.RevokeAgentIngressCredential
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: credentialId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RevokeAgentIngressCredentialResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeAgentIngressCredentialResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RevokeAgentIngressCredentialResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
  '/api/v1/agents/{agentId}/features':
    patch:
      tags:
        - Agents
      summary: Update agent features
      description: "Protection changes require Idempotency-Key and If-Match headers matching idempotencyKey and expectedRowVersion in the reviewed body. Reload and review after a version conflict. This does not change the agent lifecycle.\n\nAuthorization policy: AdministratorOnly."
      operationId: Agents.UpdateFeatures
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          description: Required for protection changes; must match body idempotencyKey (UUID v4).
          schema:
            type: string
        - name: If-Match
          in: header
          description: Required for protection changes; one exact row version matching body expectedRowVersion. Read the latest resource before reviewing a new change.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFeaturesRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateFeaturesRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateFeaturesRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UpdateFeaturesResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateFeaturesResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/UpdateFeaturesResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
  '/api/v1/agents/{agentId}:enable':
    post:
      tags:
        - Agents
      summary: Enable an agent
      description: "Changes gateway admission state for the selected registration.\n\nAuthorization policy: AdministratorOrOperator."
      operationId: Agents.EnableAgent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AgentStateChangeResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStateChangeResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AgentStateChangeResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
...
components:
  schemas:
...
    RegisterAgentRequest:
      required:
        - externalAgentId
        - name
        - description
        - ownerObjectId
        - environment
        - features
      type: object
      properties:
        externalAgentId:
          type: string
        name:
          type: string
        description:
          type:
            - 'null'
            - string
        ownerObjectId:
          type: string
        environment:
          type: string
        features:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/AgentFeaturesDto'
        blueprint:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/AgentBlueprintSelectionDto'
    RegisterAgentResponse:
      required:
        - agentId
        - externalAgentId
        - name
        - status
        - operationId
        - createdAtUtc
        - links
      type: object
      properties:
        agentId:
          type: string
          format: uuid
        externalAgentId:
          type: string
        name:
          type: string
        status:
          type: string
        operationId:
          type: string
          format: uuid
        createdAtUtc:
          type: string
          format: date-time
        links:
          type:
            - 'null'
            - object
          additionalProperties:
            type: string
        gatewayCredential:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/AgentGatewayCredentialDto'
...
  securitySchemes:
    EntraBearer:
      type: http
      description: Entra access token for the Gateway API audience with the endpoint's required gateway role. Do not paste a gateway agent key here.
      scheme: bearer
      bearerFormat: JWT
    AgentKey:
      type: http
      description: One-time gateway-issued agent key (a365gw_v1_...). Scoped to its registered agent. Do not use an Entra token for agent ingress.
      scheme: bearer
...